Add 16 testing examples for all possible axis orientations (class SVG.h|cpp)
In the program gamma-local, version 2.25, an error was introduced in
the generation of graphic output in SVG format (wrong orientation of
x,y axis).

In this directory, there is a set of examples with all sixteen
variants of eight axis orientations and two angle orientations,
calculated from the program version 2.24, which you can visually
compare with the result of the current version of the gama-local
program.